refactor: 애니메이션 스테이트 머신과 행동 트리 파일 분리
